bech32: move to common/
