+++ common/bech32.c
