+++ common/bech32.h
