+++ lightningd/bolt11.c
@@ -7,11 +7,11 @@
 #include <ccan/endian/endian.h>
 #include <ccan/structeq/structeq.h>
 #include <ccan/tal/str/str.h>
+#include <common/bech32.h>
 #include <common/utils.h>
 #include <errno.h>
 #include <hsmd/gen_hsm_wire.h>
 #include <inttypes.h>
-#include <lightningd/bech32.h>
 #include <lightningd/bolt11.h>
 #include <lightningd/hsm_control.h>
 #include <lightningd/jsonrpc.h>
